# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water Tracker App",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Water Tracker App"
  },
  "servers": [
    {
      "url": "https://project-group-05-backend.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/user": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update User's Data",
        "operationId": "updateUser",
        "description": "Partially or fully update the User data, except avatar. Avatars need to be added or updated in a separate route. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "ladygaga@gmail.com",
                    "description": "User`s email. Should be in format example@mail.com."
                  },
                  "password": {
                    "type": "string",
                    "example": "1234567890",
                    "description": "Current password required for password change."
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "0987654321",
                    "description": "New password to replace the old one."
                  },
                  "name": {
                    "type": "string",
                    "example": "Lady Gaga",
                    "description": "User`s name. Should be at most 32 characters."
                  },
                  "gender": {
                    "type": "string",
                    "example": "female",
                    "description": "User`s gender. Should be be one of the suggested values female or male. Default value is female."
                  },
                  "dailyNorm": {
                    "type": "number",
                    "example": 1500,
                    "description": "User`s daily norm of water. Should be at most 15000 ml. Default value is 2000 ml."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request successful. The server has responded as required. The server returns an object with updated User`s data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "_id",
                    "name",
                    "email",
                    "gender",
                    "dailyNorm",
                    "avatar"
                  ],
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "67a210a28f7f42ba355eb221"
                    },
                    "name": {
                      "type": "string",
                      "example": "Lady Gaga"
                    },
                    "email": {
                      "type": "string",
                      "example": "ladygaga@gmail.com"
                    },
                    "gender": {
                      "type": "string",
                      "example": "female"
                    },
                    "dailyNorm": {
                      "type": "number",
                      "example": 1500
                    },
                    "avatar": {
                      "type": "string",
                      "example": "https://res.cloudinary.com/dhfwc4nmz/image/upload/v1738667492/ysf70wtw63hihtf14spo.avif"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. The server could not understand the request. Validation error.",
            "content": {
              "application\\json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Email should be in format example@mail.com. Password should be at least 8 characters. Password should be at most 64 characters. Name should be at most 32 characters. Gender must be either \"male\" or \"female\". Daily norma should be at least 1. Daily norma should be at most 15000."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "description": "Conflict Error. The request conflicts with the current state of the server.",
            "content": {
              "application\\json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "Email in use."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "401": {
        "description": "Unauthorized error. The request is unauthenticated.",
        "content": {
          "application\\json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Not authorizated. Please provide Authorization header. Auth header should be of type Bearer."
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found error. Requested resource could not be found.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Route not found"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application\\json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Server Error. Something went wrong."
                }
              }
            }
          }
        }
      }
    }
  }
}